pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
    "discord-py>=2.5.2",
    "numpy>=2.3.2",
    "openai>=1.100.2",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "pynacl>=1.5.0",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "youtube-dl>=2021.12.17",
    "yt-dlp>=2025.8.11",
]
